refactor(PieceType): Piece의 Position을 제거하고 PieceType을 Piece로 변경

+++ src/main/java/domain/position/File.java
@@ -19,7 +19,7 @@ public class File {
     }
 
     public File add(final int movement) {
-        return new File((char) ((int) letter + movement));
+        return new File((char) (letter + movement));
     }
 
     @Override
